requirements.txt
requests
python-dotenv
pandas
scikit-learn
joblib
streamlit
